#9 Feat-[컴포넌트 구조 수정 완료]
