+++ components/mypage/Schedule.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Schedule = () => {
+  return (
+    <div>Schedule</div>
+  )
+}
+
+export default Schedule
