--- components/mypage/UserStatusView.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-
-// 좌측 유저 상태 부분
-const UserStatusView = () => {
-  return (
-    <div>UserStatusView</div>
-  )
-}
-
-export default UserStatusView
